refactor(history): move history title generation functions to history.go
