+++ completion.go
@@ -7,7 +7,6 @@ import (
 	"io"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"slices"
 	"strings"
 	"syscall"
@@ -273,6 +272,10 @@ func (s *CompletionService) runOneShotCompletionStreaming(ctx context.Context, r
 	if err := s.saveHistory(); err != nil {
 		return fmt.Errorf("failed to save history: %w", err)
 	}
+
+	if err := s.renameChatHistory(ctx); err != nil {
+		return fmt.Errorf("failed to rename history: %w", err)
+	}
 	return nil
 }
 
@@ -291,6 +294,9 @@ func (s *CompletionService) runOneShotCompletion(ctx context.Context, runCfg Run
 	runCfg.Stdout.Write([]byte(response))
 	if err := s.saveHistory(); err != nil {
 		return fmt.Errorf("failed to save history: %w", err)
+	}
+	if err := s.renameChatHistory(ctx); err != nil {
+		return fmt.Errorf("failed to rename history: %w", err)
 	}
 	return nil
 }
@@ -449,71 +455,4 @@ func (s *CompletionService) generateResponse(ctx context.Context, runCfg RunOpti
 // Whitespace is trimmed from the end of the message.
 func (s *CompletionService) SetNextCompletionPrefill(content string) {
 	s.nextCompletionPrefill = strings.TrimRight(content, " \t\n")
-}
-
-// generateHistoryTitle sends the conversation history to the LLM to generate a descriptive title
-func (s *CompletionService) generateHistoryTitle(ctx context.Context) (string, error) {
-	// Don't try to generate a title if we have no messages
-	if len(s.payload.Messages) < 2 {
-		return "empty-chat", nil
-	}
-
-	prompt := "Generate a kebab case title for the following conversation. An example is debug-rust-code or explain-quantum-mechanics."
-	msgLimit := min(len(s.payload.Messages), 10)
-	for _, m := range s.payload.Messages[:msgLimit] {
-		for _, p := range m.Parts {
-			prompt += fmt.Sprint(p)
-		}
-	}
-
-	completion, err := llms.GenerateFromSinglePrompt(ctx, s.model, prompt)
-	if err != nil {
-		return "", fmt.Errorf("failed to generate title: %w", err)
-	}
-
-	fmt.Println("completion", completion)
-
-	// If title is too long, truncate it
-	const maxTitleLength = 50
-	if len(completion) > maxTitleLength {
-		completion = completion[:maxTitleLength]
-	}
-
-	return completion, nil
-}
-
-// renameChatHistory generates a title and renames the history file
-func (s *CompletionService) renameChatHistory(ctx context.Context) error {
-	if s.disableHistory {
-		return nil
-	}
-	if s.historyOutFile == "" {
-		home, err := os.UserHomeDir()
-		if err != nil {
-			return fmt.Errorf("failed to get user home directory: %w", err)
-		}
-
-		// Get the current history file path
-		currentPath := filepath.Join(home, ".cgpt", fmt.Sprintf("default-history-%s.yaml", s.sessionTimestamp))
-
-		// Generate a descriptive title
-		title, err := s.generateHistoryTitle(ctx)
-		if err != nil {
-			return fmt.Errorf("failed to generate title: %w", err)
-		}
-
-		// Create new filename with timestamp + title
-		newPath := filepath.Join(home, ".cgpt", fmt.Sprintf("%s.yaml", title))
-
-		// Rename the file
-		if err := os.Rename(currentPath, newPath); err != nil {
-			return fmt.Errorf("failed to rename history file: %w", err)
-		}
-
-		fmt.Fprintf(s.Stderr, "\033[38;5;240mcgpt: Renamed history to: %s\033[0m\n", filepath.Base(newPath))
-
-		// Update the historyOutFile to use the new path
-		s.historyOutFile = newPath
-	}
-	return nil
 }
